Build 5 patch 1, TensorRT, README edits, results organization, images

+++ Tensorflow_TensorRT/AIXPRT/Harness/hddlSetup.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
-echo "Loading HDDL dependencies"
-sudo kill -9 $( pidof hddldaemon )
-sudo rm -rf /var/tmp/hddl_*
-source /opt/intel/computer_vision_sdk/bin/setupvars.sh
-/opt/intel/computer_vision_sdk/deployment_tools/inference_engine/external/hddl/bin/hddldaemon
+
+if [ ! $( pidof hddldaemon ) ]; then
+
+  echo "Loading HDDL dependencies"
+  sudo rm -rf /var/tmp/hddl_*
+  source /opt/intel/computer_vision_sdk/bin/setupvars.sh
+  /opt/intel/computer_vision_sdk/deployment_tools/inference_engine/external/hddl/bin/hddldaemon
+
+fi
